+++ src/MeasureMap/ProfilerResultCollectionExtensions.cs
@@ -20,8 +20,8 @@ namespace MeasureMap
             sb.AppendLine("## MeasureMap Benchmark");
             sb.AppendLine($" Iterations:\t\t{results.Iterations}");
             sb.AppendLine($"### Summary");
-            sb.AppendLine("| Name | Avg Time | Avg ms | Avg Ticks | Total | Fastest | Slowest | Iterations |");
-            sb.AppendLine("|--- |---: |---: |---: |---: |---: |---: |---: |");
+            sb.AppendLine("| Name | Avg Time | Avg ms | Avg Ticks | Total | Fastest | Slowest | Iterations | Throughput |");
+            sb.AppendLine("|--- |---: |---: |---: |---: |---: |---: |---: |---: |");
             foreach (var key in results.Keys)
             {
                 sb.AppendLine($"| {key} {TraceLine(results[key])}");
@@ -36,7 +36,7 @@ namespace MeasureMap
 
         private static string TraceLine(IProfilerResult result)
         {
-            return $"| {result.AverageTime} | {result.AverageMilliseconds} | {result.AverageTicks} | {result.TotalTime.ToString()} | {result.Fastest.Ticks} | {result.Slowest.Ticks} | {result.Iterations.Count()} |";
+            return $"| {result.AverageTime} | {result.AverageMilliseconds} | {result.AverageTicks} | {result.TotalTime.ToString()} | {result.Fastest.Ticks} | {result.Slowest.Ticks} | {result.Iterations.Count()} | {result.Throughput()} |";
         }
     }
 }